# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lumen — AI-powered wellness companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile, condition mode, premium status
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile (name, condition mode, daily targets)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/upgrade:
    post:
      operationId: togglePremium
      tags: [profile]
      summary: Toggle premium membership (mock freemium upgrade)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                premium:
                  type: boolean
              required: [premium]
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
  /workouts:
    get:
      operationId: listWorkouts
      tags: [workouts]
      summary: List workouts
      parameters:
        - in: query
          name: date
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
    post:
      operationId: createWorkout
      tags: [workouts]
      summary: Log a workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
  /workouts/{id}:
    delete:
      operationId: deleteWorkout
      tags: [workouts]
      summary: Delete a workout
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List reminders (AI-driven + user-created)
      responses:
        "200":
          description: Reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    patch:
      operationId: updateReminder
      tags: [reminders]
      summary: Toggle / edit reminder
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReminderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete reminder
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /planner/weekly:
    get:
      operationId: getWeeklyPlan
      tags: [planner]
      summary: "Weekly plan: per-day suggested actions, condition-tailored"
      parameters:
        - in: query
          name: weekStart
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Weekly plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"
  /planner/monthly:
    get:
      operationId: getMonthlyPlan
      tags: [planner]
      summary: Monthly plan with weekly themes and goals
      parameters:
        - in: query
          name: month
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Monthly plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyPlan"
...
  /vision/food:
    post:
      operationId: analyzeFoodPhoto
      tags: [vision]
      summary: Analyze a food photo and return detected items + nutrition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeFoodPhotoBody"
      responses:
        "200":
          description: Detected food and nutrition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodAnalysisResult"
...
components:
  schemas:
...
    HealthMode:
      type: string
      enum: [standard, diabetes, hypertension, heart_health, pregnancy, weight_loss]
      description: Personalized mode that adapts AI recommendations
    Profile:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        avatarColor: { type: string, description: "Tailwind-compatible accent for avatar" }
        mode: { $ref: "#/components/schemas/HealthMode" }
        premium: { type: boolean }
        dailyCalorieTarget: { type: integer }
        dailyProteinTarget: { type: integer }
        dailySleepTargetHours: { type: number }
        dailyStepsTarget: { type: integer }
        dailyScreenTimeLimitMinutes: { type: integer }
        glucoseTargetLow: { type: integer }
        glucoseTargetHigh: { type: integer }
        dailyCarbLimit: { type: integer }
        onboardingComplete: { type: boolean }
        voiceEnabled: { type: boolean }
        motionPermissionGranted: { type: boolean }
        notificationsEnabled: { type: boolean }
        joinedAt: { type: string, format: date-time }
      required: [id, name, mode, premium, dailyCalorieTarget, dailyProteinTarget, dailySleepTargetHours, dailyStepsTarget, dailyScreenTimeLimitMinutes, glucoseTargetLow, glucoseTargetHigh, dailyCarbLimit, onboardingComplete, voiceEnabled, motionPermissionGranted, notificationsEnabled, joinedAt]
    UpdateProfileBody:
      type: object
      properties:
        name: { type: string }
        mode: { $ref: "#/components/schemas/HealthMode" }
        dailyCalorieTarget: { type: integer }
        dailyProteinTarget: { type: integer }
        dailySleepTargetHours: { type: number }
        dailyStepsTarget: { type: integer }
        dailyScreenTimeLimitMinutes: { type: integer }
        glucoseTargetLow: { type: integer }
        glucoseTargetHigh: { type: integer }
        dailyCarbLimit: { type: integer }
        voiceEnabled: { type: boolean }
        motionPermissionGranted: { type: boolean }
        notificationsEnabled: { type: boolean }
...
    Workout:
      type: object
      properties:
        id: { type: string }
        type:
          type: string
          enum: [walk, run, cycle, strength, yoga, swim, hiit, sport, other]
        durationMinutes: { type: integer }
        caloriesBurned: { type: integer }
        steps: { type: integer }
        intensity:
          type: string
          enum: [light, moderate, vigorous]
        loggedAt: { type: string, format: date-time }
        source:
          type: string
          enum: [manual, sensor]
      required: [id, type, durationMinutes, caloriesBurned, steps, intensity, loggedAt, source]
    CreateWorkoutBody:
      type: object
      properties:
        type:
          type: string
          enum: [walk, run, cycle, strength, yoga, swim, hiit, sport, other]
        durationMinutes: { type: integer }
        caloriesBurned: { type: integer }
        steps: { type: integer }
        intensity:
          type: string
          enum: [light, moderate, vigorous]
        source:
          type: string
          enum: [manual, sensor]
      required: [type, durationMinutes, caloriesBurned, steps, intensity, source]
...
    Reminder:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        time: { type: string, description: "HH:MM 24h" }
        category:
          type: string
          enum: [hydration, meal, movement, sleep, screen, mindfulness, medication]
        enabled: { type: boolean }
        repeatDays:
          type: array
          items:
            type: string
            enum: [mon, tue, wed, thu, fri, sat, sun]
        aiGenerated: { type: boolean }
      required: [id, title, time, category, enabled, repeatDays, aiGenerated]
    CreateReminderBody:
      type: object
      properties:
        title: { type: string }
        time: { type: string }
        category:
          type: string
          enum: [hydration, meal, movement, sleep, screen, mindfulness, medication]
        enabled: { type: boolean }
        repeatDays:
          type: array
          items:
            type: string
            enum: [mon, tue, wed, thu, fri, sat, sun]
      required: [title, time, category, enabled, repeatDays]
    UpdateReminderBody:
      type: object
      properties:
        title: { type: string }
        time: { type: string }
        category:
          type: string
          enum: [hydration, meal, movement, sleep, screen, mindfulness, medication]
        enabled: { type: boolean }
        repeatDays:
          type: array
          items:
            type: string
            enum: [mon, tue, wed, thu, fri, sat, sun]
...
    PlannedAction:
      type: object
      properties:
        category:
          type: string
          enum: [nutrition, sleep, activity, screen, mindfulness, condition]
        title: { type: string }
        detail: { type: string }
        targetValue:
          type: ["string", "null"]
      required: [category, title, detail]
    WeeklyPlanDay:
      type: object
      properties:
        date: { type: string }
        weekday: { type: string }
        focus: { type: string }
        actions:
          type: array
          items:
            $ref: "#/components/schemas/PlannedAction"
      required: [date, weekday, focus, actions]
    WeeklyPlan:
      type: object
      properties:
        weekStart: { type: string }
        weekEnd: { type: string }
        modeContext: { type: string, description: "Health mode the plan is tailored to" }
        theme: { type: string }
        days:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyPlanDay"
      required: [weekStart, weekEnd, modeContext, theme, days]
    MonthlyPlanWeek:
      type: object
      properties:
        weekLabel: { type: string }
        weekStart: { type: string }
        theme: { type: string }
        goals:
          type: array
          items: { type: string }
      required: [weekLabel, weekStart, theme, goals]
    MonthlyPlan:
      type: object
      properties:
        month: { type: string }
        modeContext: { type: string }
        narrative: { type: string }
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyPlanWeek"
      required: [month, modeContext, narrative, weeks]
...
    AnalyzeFoodPhotoBody:
      type: object
      properties:
        imageBase64:
          type: string
          description: Image data URL or raw base64
        hint:
          type: string
          description: Optional textual hint about what's in the photo
      required: [imageBase64]
    DetectedFoodItem:
      type: object
      properties:
        name: { type: string }
        portion: { type: string }
        calories: { type: integer }
        proteinGrams: { type: integer }
        carbsGrams: { type: integer }
        fatGrams: { type: integer }
        confidence: { type: number, minimum: 0, maximum: 1 }
      required: [name, portion, calories, proteinGrams, carbsGrams, fatGrams, confidence]
    FoodAnalysisResult:
      type: object
      properties:
        mealName: { type: string }
        suggestedMealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        items:
          type: array
          items: { $ref: "#/components/schemas/DetectedFoodItem" }
        totalCalories: { type: integer }
        totalProteinGrams: { type: integer }
        totalCarbsGrams: { type: integer }
        totalFatGrams: { type: integer }
        modelNotes: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
      required: [mealName, suggestedMealType, items, totalCalories, totalProteinGrams, totalCarbsGrams, totalFatGrams, confidence]
...